+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [FIXED] - use `promise.then` rather than `promise.return` method
+
 ## v1.0.0 - initial release
 
 * [ADDED] - `EventEmitter.prototype.emitAsync` & `emitAsyncSeries` & `emitAsyncParallel` methods